+++ src/main/java/com/spring/spring_booking_system/repositories/SpaceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpaceRepository extends CrudRepository<Space, Integer> {
+public interface SpaceRepository extends CrudRepository<Space, Long> {
 }
